+++ base/src/AzCore/Math/RandomNumberGenerator.cpp
@@ -40,7 +40,7 @@ void RandomNumberGenerator::Seed(u64 seed)
 f32 random(f32 min, f32 max, RandomNumberGenerator &rng)
 {
     u32 num = rng.Generate();
-    return ((f64)num * (f64)(max - min) / (f64)UINT32_MAX) + (f64)min;
+    return (f32)(((f64)num * (f64)(max - min) / (f64)UINT32_MAX) + (f64)min);
 }
 
 i32 random(i32 min, i32 max, RandomNumberGenerator &rng)
